### 2017-06-08 #286 @whedon
[PRE REVIEW]: Launcher: A simple tool for executing high throughput computing workloads
**Submitting author:** @lwilson (<a href="http://orcid.org/0000-0002-0686-3152">Lucas Wilson</a>)
**Repository:** <a href="https://github.com/TACC/launcher" target ="_blank">https://github.com/TACC/launcher</a>
**Version:** v3.1
