### 2017-06-08 #286 @danielskatz
@kc9qey - would you be willing to review this submission for JOSS?
