### 2017-06-08 #286 @kc9qey
Yeah, only because I can :

===============================================================================
                          --- Stampede 2 is coming
---
     --- Stampede 1 will start phased decommissioning in January 2017
---
          --- The KNLs are now available <> Start migrating today
---
            https://portal.tacc.utexas.edu/user-news/-/news/101909

            https://portal.tacc.utexas.edu/user-guides/stampede#knl

===============================================================================
login3.stampede(1)$ id
uid=456924(tg456924) gid=80228(G-80228)
groups=80228(G-80228),80871(G-80871),80898(G-80898),81815(G-81815),802826(G-802826),803434(G-803434),803442(G-803442),803452(G-803452),803673(G-803673),815561(G-815561),816467(G-816467),816890(G-816890),818338(G-818338)
login3.stampede(2)$

On Thu, Jun 8, 2017 at 8:19 AM, Daniel S. Katz <notifications@github.com>
wrote:

> @kc9qey <https://github.com/kc9qey> - would you be willing to review this
> submission for JOSS?
>
