### 2017-06-08 #286 @danielskatz
tell me about how I phrase the "secret" banana again?
