### 2017-06-08 TACC/launcher#18 @kc9qey
pre-requisite needs mentioned
Launcher needs python 2.7 or higher ? , not mentioned in install or readme
