### 2017-06-08 TACC/launcher#19 @kc9qey
missing example usage
A full example usage is missing so it's difficult to know what to do.  I might piece it together after staring at it for an hour but that's too high a bar for something that is trying to appear simple and easy to use.
