### 2017-06-08 TACC/launcher#20 @kc9qey
Along with example usage, missing automated test(s). 
An automated test would go a long way toward meeting the example usage.
