### 2017-06-08 TACC/launcher#18 @lwilson
I have updated both INSTALL and README.md to mention the Python 2.7+ requirement, as well as more prominently note the hwloc requirement for automatic task binding.
