### 2017-06-08 #289 @danielskatz
just adding links to make it easier to follow the previous comment:

> TACC/launcher Functionality documentation needed.
> https://github.com/TACC/launcher/issues/21 opened a minute ago by kc9qey
> TACC/launcher Along with example usage, missing automated test(s).
> https://github.com/TACC/launcher/issues/20 opened 2 minutes ago by kc9qey
> TACC/launcher missing example usage
> https://github.com/TACC/launcher/issues/19 opened 3 minutes ago by kc9qey
> TACC/launcher pre-requisite needs mentioned
> https://github.com/TACC/launcher/issues/18 opened 5 minutes ago by kc9qey
