### 2017-06-08 TACC/launcher#21 @kc9qey
Launcher

Launcher is a utility for performing simple, data parallel, high throughput computing (HTC) workflows on clusters, massively parallel processor (MPP) systems, workgroups of computers, and personal machines.

...something more than that.  What are some use cases?  What does a typical workflow look like (diagram ? )
