### 2017-06-11 #289 @danielskatz
Hi @lwilson - It looks like you've resolves one of the four issues above, but still need to work on the other three.  I'm just checking on the status of them, so it's clear who has the ball currently.
