### 2017-06-26 #289 @danielskatz
@lwilson - the review process is waiting on your resolution of the last 3 issues created by @kc9qey
