### 2017-07-17 #289 @lwilson
Dan and Galen,

I’m still alive! Sorry for the delay on this – I’ll be pushing changes and updating those tickets this week.

________________________________
Lucas A. Wilson, Ph.D.
Director, Training and Professional Development
Texas Advanced Computing Center
The University of Texas at Austin
ACB 2.278
+1 512 232 7351
lwilson@tacc.utexas.edu<mailto:lwilson@tacc.utexas.edu>
