### 2017-07-18 291f8aa
Including section on how to cite launcher
