### 2017-07-18 44cb4a7
Fixing reference to paper.bib
