+++ README.md
@@ -62,4 +62,4 @@ The directory containing this README contains several example submission scripts
   * SLURM: launcher.slurm
 
 ## Referencing Launcher
-If you are using Launcher, please remember to make a reference to it when publishing results. The file `paper/launcher.bib` contains the BibTeX-formatted citation list. Please reference entry `launcher` (i.e., in LaTeX: `\cite{Wilson:2014:LSF:2616498.2616534}`).
+If you are using Launcher, please remember to make a reference to it when publishing results. The file `paper/paper.bib` contains the BibTeX-formatted citation list. Please reference entry `launcher` (i.e., in LaTeX: `\cite{Wilson:2014:LSF:2616498.2616534}`).
