### 2017-07-18 09c87ef
Fixing bibtex reference
