### 2017-07-25 #289 @danielskatz
@lwilson - not sure what "this" week meant... :)
