### 2017-07-25 TACC/launcher#19 @lwilson
Interactive walk through in process in tests/ (see  #23 and #26). Script acts as a tutorial/example.
