### 2017-07-25 b836462
Updating paper with use cases and workflow figure
