### 2017-07-25 TACC/launcher#21 @lwilson
#27 Updates the paper with use cases and a figure which describes launcher's workflow. Paper has also been updated with names of other project contributors.
