### 2017-07-25 #289 @lwilson
It's apparently a "baker's" week (it has 1 extra day)!
