### 2017-07-25 5963225
Added descrptions of enviromental variables
