+++ README.md
@@ -1,6 +1,5 @@
 # Launcher
 Launcher is a utility for performing simple, data parallel, high throughput computing (HTC) workflows on clusters, massively parallel processor (MPP) systems, workgroups of computers, and personal machines.
-this i s a test
 ## Installing Launcher
 Launcher does not need to be compiled. Unpack the tarball or clone the repository in the desired directory. Then, set `LAUNCHER_DIR` to point to that location. Python 2.7 or greater and hwloc are required for full functionality. See INSTALL for more information.
 
@@ -27,15 +26,21 @@ The launcher defines the following environment variables for each job that is st
 * `$LAUNCHER_NJOBS` contains the number of jobs in your job file.
 * `$LAUNCHER_TSK_ID` is the particular processing core that the job is running on, from 0 to `$LAUNCHER_NPROCS-1`.
 * `$LAUNCHER_JID` represents the particular job instance currently running. `$LAUNCHER_JID` is numbered from 1 to `$LAUNCHER_NJOBS`
-$LAUNCHER_JOB_FILE'
-the launcher job file is the specific directory where the specified output statement(s) is/are. Here you will find the file with the jobs that will run in the your 'parametric submission'.
-.
 
 Example: If you want to redirect stdout to a file containing the unique ID of each line, you can specify the following in the paramlist file: ```a.out > out.o$LAUNCHER_JID```
 
 If this particular execution instance of a.out was the first line in the job file, the output would be placed in the file "out.o1".
 
 Note: you can also use the launcher to run a sequence of serial jobs when you have more jobs to run than the requested number of processors.  
+## Enviromental Variable descriptions
+* '$LAUNCHER_JOB_FILE' - The launcher job file is the specific directory where the specified output statement(s) is/are. Here you will find the file with the jobs that will run in the your 'parametric submission'.
+* 'LAUNCHER_WORKDIR' - This is simply the working directory. This directory will hold any relative path as well as the execution of launcher. In most cases, the working directory will be your home directory or your root users home directory.
+* 'LAUNCHER_NPROCS' - In your parametric submissions, there are processes that are running. This variable stores the number of processes that are simultaneously running in your submissions . It can simply be stated as the total processes. In this example, when keeping variables their default vales, the value will be 1.
+* '$LAUNCHER_NHOSTS' - Just as launcher_nprocs holds the number of processes running simultaneously, this variable contains the number of hosts that are running in your submission simultaneously.
+* '$LAUNCHER_PPN' - This is the number of processors used while running the program. Changing the number of processors is an easy way to increase or decrease the time spent running the program. The default for this variable is 1.
+* '$LAUNCHER_NJOBS' - This variable is simply the total number of jobs. If only one job is being run, then the variable value is 1. 
+* '$LAUNCHER_TSK_ID' - When running the program, this variable is a number from 0 to 'launcher_nprocs-1' which represents the processing core on which the job is running.
+* '$LAUNCHER_JID' - Numbered from 1 to the variable 'launcher_njobs', this variable is an active representation of the job instance currently running in the problem.
 
 ## Task Scheduling Behavior
 
@@ -67,4 +72,4 @@ The directory containing this README contains several example submission scripts
   * SLURM: launcher.slurm
 
 ## Referencing Launcher
-If you are using Launcher, please remember to make a reference to it when publishing results. The file `paper/paper.bib` contains the BibTeX-formatted citation list. Please reference entry `Wilson:2014:LSF:2616498.2616533` (i.e., in LaTeX: `\cite{Wilson:2014:LSF:2616498.2616534}`).
[…]
