### 2017-07-25 TACC/launcher#19 @kc9qey
Lucas,

Can you send me a transcript of the script being run.  I'm unable to get it
to work properly, and frequently get:

Please do so again: you did not enter an appropiate y or n. Please do so
again: you did not enter an appropiate y or n. Please do so again: you did
not enter an appropiate y or n. Please do so again: you did not enter an
appropiate y or n. Please do so again: you did not enter an appropiate y or
n. Please do so again: you did not enter an appropiate y or n. Please do so
again: you did not enter an appropiate y or n. Please do so again: you did
not enter an appropiate y or n. Please do so again: you did not enter an
appropiate y or n. Please do so again: you did not enter an appropiate y or
n. Please do so again: you did not enter an appropiate y or n. Please do so
again: you did not enter an appropiate y or n. Please do so again: you did
not enter an appropiate y or n. Please do so again: you did not enter an
appropiate y or n. Please do so again: you did not enter an appropiate y or
n. Please do so again: you did not enter an appropiate y or n. Please do so
again: you did not enter an appropiate y or n. Please do so again: you did
not enter an appropiate y or n. Please do so again: you did not

On Tue, Jul 25, 2017 at 9:33 AM, Lucas A. Wilson <notifications@github.com>
wrote:

> Interactive walk through in process in tests/ (see #23
> <https://github.com/TACC/launcher/pull/23> and #26
> <https://github.com/TACC/launcher/pull/26>). Script acts as a
> tutorial/example.
>
