### 2017-07-25 TACC/launcher#19 @lwilson
I've put a new script quicktest in the /tests directory which will allow you to do a sanity test. No inputs required on your end, just run:
`cd tests; ./quicktest`

I'm going to have @marshalllerner debug the interactive walkthrough.
