### 2017-07-25 TACC/launcher#19 @kc9qey
cool, looks good now

On Tue, Jul 25, 2017 at 2:58 PM, Lucas A. Wilson <notifications@github.com>
wrote:

> I've put a new script quicktest in the /tests directory which will allow
> you to do a sanity test. No inputs required on your end, just run:
> cd tests; ./quicktest
>
> I'm going to have @marshalllerner <https://github.com/marshalllerner>
> debug the interactive walkthrough.
>
