### 2017-07-25 TACC/launcher#19 @marshalllerner
for some reason when I correct something and then save it it adds the /r carriage returns because I'm using Windows OS. I'll correct that now.
