### 2017-07-26 TACC/launcher#19 @lwilson
#31 adds a section to the README.md about verifying Launcher installation using the quicktest script.
