+++ .travis.yml
@@ -1,0 +1,11 @@
+language: python
+python: 
+    - "2.7"
+    - "2.7"
+    - "3.2"
+    - "3.3"
+    - "3.4"
+    - "3.5"
+
+install: "pip install -r requirements.txt"
+script: py.test
