+++ .travis.yml
@@ -7,5 +7,5 @@ python:
     - "3.4"
     - "3.5"
 
-install: "pip install -r requirements.txt"
-script: py.test
+# install: "pip install -r requirements.txt"
+# script: py.test
