+++ .travis.yml
@@ -1,4 +1,8 @@
 language: bash
 
+os:
+  - linux
+  - osx
+
 script:
     - bash .travis_test_file
