### 2017-07-31 TACC/launcher#20 @lwilson
The quicktest script has been integrated with Travis-CI to do automated tested on the repo from both linux and osx environments. build status button has been added to README.md (#35)
