### 2017-08-01 #289 @lwilson
We have made some updates on:
https://github.com/TACC/launcher/issues/21
https://github.com/TACC/launcher/issues/20
https://github.com/TACC/launcher/issues/19
