### 2017-08-01 #289 @danielskatz
@kc9qey - back to you...
