### 2017-08-01 18c56bf
made a function (and exception DynamicExeException in case a dynamic executable file contains more than just standard libraries) in Python that should be able to be given a direct or relative path to a job file and return true if the file is possible for launcher and false otherwise
