### 2017-08-01 fb8bf9f
added a return statement to the case in which a dynamic executable file contains more than just standard libraries (it should raise an exception beforehand but just in case)
