### 2017-08-02 94ecb9b
added a file that given a job file splits the commands into two files, one for BOINC (BOINCcommands.txt) and one for launcher (launchercommands.txt)
