### 2017-08-02 c60cf93
added some failcases and a small test for the edge cases of those tests
