### 2017-08-07 #289 @danielskatz
@kc9qey - just another ping ...
