### 2017-08-07 #289 @kc9qey
I don't have anything else to review on this.  They hit all my points I
think.  I was in the hospital last week getting an important upgrade.

On Mon, Aug 7, 2017 at 7:44 AM, Daniel S. Katz <notifications@github.com>
wrote:

> @kc9qey <https://github.com/kc9qey> - just another ping ...
>
