### 2017-08-07 #289 @danielskatz
@kc9qey - there are still a couple of boxes in the checklist that are not checked.
