### 2017-08-07 #289 @kc9qey
I didn't see any automated tests, so left unchecked.
Performance claims for me are a "won't do" --also left unchecked.
