### 2017-08-07 #289 @danielskatz
@kc9qey

If there are no performance claims, you can check that box.

Are you also saying that there are no automated tests or manual steps described so that the function of the software can be verified?  If so, then this box cannot be checked, and we need to go back to @lwilson
