### 2017-08-07 #289 @kc9qey
"manual steps described so that the function of the software can be verified" , i can take that for the automated tests.
