### 2017-08-11 #289 @arfon
@kc9qey many thanks for your review and thanks to @danielskatz for editing this submission ✨

@lwilson - your paper is now accepted into JOSS. Your DOI is http://dx.doi.org/10.21105/joss.00289 ⚡️ 🚀 💥
